errors: move errors module to transport module
Originally, all errors could be found under scylla::transport::errors.
Let's retain the same path.
