+++ scylla/src/transport/mod.rs
@@ -3,6 +3,7 @@ mod cluster;
 pub(crate) mod connection;
 mod connection_pool;
 pub mod downgrading_consistency_retry_policy;
+pub mod errors;
 pub mod execution_profile;
 pub mod host_filter;
 pub mod iterator;
