+++ scylla/src/transport/retry_policy.rs
@@ -2,8 +2,8 @@
 //! To decide when to retry a query the `Session` can use any object which implements
 //! the `RetryPolicy` trait
 
-use crate::errors::{DbError, QueryError, WriteType};
 use crate::frame::types::Consistency;
+use crate::transport::errors::{DbError, QueryError, WriteType};
 
 /// Information about a failed query
 pub struct QueryInfo<'a> {
@@ -219,10 +219,10 @@ impl RetrySession for DefaultRetrySession {
 #[cfg(test)]
 mod tests {
     use super::{DefaultRetryPolicy, QueryInfo, RetryDecision, RetryPolicy};
-    use crate::errors::{BadQuery, QueryError};
-    use crate::errors::{DbError, WriteType};
     use crate::statement::Consistency;
     use crate::test_utils::setup_tracing;
+    use crate::transport::errors::{BadQuery, QueryError};
+    use crate::transport::errors::{DbError, WriteType};
     use bytes::Bytes;
     use std::io::ErrorKind;
     use std::sync::Arc;
