+++ scylla/src/transport/session_test.rs
@@ -1,6 +1,5 @@
 use crate as scylla;
 use crate::batch::{Batch, BatchStatement};
-use crate::errors::{BadKeyspaceName, BadQuery, DbError, QueryError};
 use crate::frame::response::result::Row;
 use crate::prepared_statement::PreparedStatement;
 use crate::query::Query;
@@ -10,6 +9,7 @@ use crate::statement::Consistency;
 use crate::test_utils::{scylla_supports_tablets, setup_tracing};
 use crate::tracing::TracingInfo;
 use crate::transport::cluster::Datacenter;
+use crate::transport::errors::{BadKeyspaceName, BadQuery, DbError, QueryError};
 use crate::transport::partitioner::{
     calculate_token_for_partition_key, Murmur3Partitioner, Partitioner, PartitionerName,
 };
@@ -2526,7 +2526,7 @@ async fn test_rate_limit_exceeded_exception() {
         }
     }
 
-    use crate::errors::OperationType;
+    use crate::transport::errors::OperationType;
 
     match maybe_err.expect("Rate limit error didn't occur") {
         QueryError::DbError(DbError::RateLimitReached { op_type, .. }, _) => {
